# src/Clients/onlineshop/src/clients/authService/specs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "UserService.Api",
    "version": "1.0"
  },
  "paths": {
    "/account/register": {
      "post": {
        "tags": ["Account"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUser"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUser"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RegisterUser": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "password": {
            "tokenUrl": "https://localhost:7101/connect/token",
            "scopes": {
              "user_api": "user api scope"
            }
          },
          "authorizationCode": {
            "authorizationUrl": "https://localhost:7101/connect/authorize",
            "tokenUrl": "https://localhost:7101/connect/token",
            "scopes": {
              "user_api": "user api scope"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "oauth2": []
    }
  ]
}